Refactored domain classes Requests extend ZonemasterRequest Responses extend ZonemasterResponse
