+++ whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/domain/StartDomainTestResponse.java
@@ -1,30 +1,19 @@
 package net.ripe.db.whois.update.dns.zonemaster.domain;
 
 
+import com.google.common.base.MoreObjects;
+
 import javax.annotation.Nullable;
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlRootElement;
 
-public class StartDomainTestResponse {
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlRootElement(name = "response")
+public class StartDomainTestResponse extends ZonemasterResponse {
 
-    private String jsonrpc;
-    private String id;
     private String result;
     private Error error;
-
-    public String getJsonrpc() {
-        return jsonrpc;
-    }
-
-    public void setJsonrpc(final String jsonrpc) {
-        this.jsonrpc = jsonrpc;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(final String id) {
-        this.id = id;
-    }
 
     /**
      * Percentage complete for Start Domain Test in progress
@@ -34,17 +23,28 @@ public class StartDomainTestResponse {
         return result;
     }
 
-    public void setResult(final String result) {
-        this.result = result;
-    }
-
     @Nullable
     public Error getError() {
         return error;
     }
 
-    static class Error {
+    @Override
+    protected MoreObjects.ToStringHelper toStringHelper() {
+        return super.toStringHelper()
+                .add("result", result)
+                .add("error", error);
+    }
+
+    public static class Error {
         private String message;
         private int code;
+
+        @Override
+        public String toString() {
+            return MoreObjects.toStringHelper(this)
+                .add("message", message)
+                .add("code", code)
+                .toString();
+        }
     }
 }
