+++ whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/domain/TestProgressRequest.java
@@ -5,19 +5,12 @@ package net.ripe.db.whois.update.dns.zonemaster.domain;
  *
  * @see <a href="https://github.com/dotse/zonemaster-backend/blob/master/docs/API.md">Zonemaster documentation</a>
  */
-public class TestProgressRequest implements ZonemasterRequest {
-
-    final Request request;
+public class TestProgressRequest extends ZonemasterRequest {
 
     public TestProgressRequest(final String id) {
-        this.request = new Request();
-        this.request.setMethod(Request.Method.TEST_PROGRESS);
-        final Request.Params params = new Request.Params();
+        super.setMethod(ZonemasterRequest.Method.TEST_PROGRESS);
+        final ZonemasterRequest.Params params = new ZonemasterRequest.Params();
         params.setId(id);
-    }
-
-    @Override
-    public Request getRequest() {
-        return request;
+        super.setParams(params);
     }
 }
