+++ whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/domain/VersionInfoRequest.java
@@ -5,18 +5,9 @@ package net.ripe.db.whois.update.dns.zonemaster.domain;
  *
  * @see <a href="https://github.com/dotse/zonemaster-backend/blob/master/docs/API.md">Zonemaster documentation</a>
  */
-class VersionInfoRequest implements ZonemasterRequest {
+public class VersionInfoRequest extends ZonemasterRequest {
 
-    private final Request request;
-
-    VersionInfoRequest() {
-        this.request = new Request();
-        this.request.setMethod(Request.Method.VERSION_INFO);
+    public VersionInfoRequest() {
+        setMethod(ZonemasterRequest.Method.VERSION_INFO);
     }
-
-    @Override
-    public Request getRequest() {
-        return this.request;
-    }
-
 }
